Create Token XML Updater
Python script that can compare new token sets on Scryfall to the current token XML. Appends token reprints to their respective entries and creates a new XML for new tokens. Does not handle reverse related, adding 'token' to most but not all tokens, or adding spaces to same-name tokens. Currently provided as a .py for running in an IDE; plan to make a cmd line version soon.