Update token links to cards.scryfall.io
Fix #201. This changes the links from the old CND on c1.scryfall.com to
links to the new CDN on cards.scryfall.io.

The changes are automated: the refresh.py script can read a tokens.xml
file containing CDN picURLs of either version, and makes requests to the
Scryfall API to generate the most up-to-date URLs for each token.

The script can be run with:

  $ python3 refresh.py -i tokens.xml
  $ python3 refresh.py -i challenge_tokens.xml

This avoids using api.scryfall.com URLs for the tokens (which are
rate-limited) while still being able to update the file with new URLs by
running the script.